+++ resources/views/todos/show.blade.php
@@ -18,7 +18,7 @@
         </div>
 
         <a href="/todos/{{ $todo->id }}/edit" class="btn btn-info my-2">Edit</a>
-        <a href="/todos/{{ $todo->id }}/delete" class="btn btn-danger my-2">Delete</a>
+        <a href="/todos/{{ $todo->id }}/delete" onclick="return confirm('Are you sure you want to delete this item?');" class="btn btn-danger my-2">Delete</a>
     </div>
 </div>
 @endsection
